requirements.txt
adal==1.2.1
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
botbuilder-core==4.11.0
botbuilder-schema==4.11.0
botframework-connector==4.11.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==2.8
idna==2.10
isodate==0.6.0
jsonpickle==1.2
msal==1.2.0
msrest==0.6.10
multidict==5.1.0
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.5.3
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.25.11
yarl==1.6.3
